pyproject.toml
[project]
name = "lexis"
version = "0.1.0"
description = "🗣️ A human-like question-answering Discord bot."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
]
